# docs/swagger.yaml
basePath: /api/v1
definitions:
  hexagonal-go-backend_internal_platform_http_response.Envelope:
    properties:
      data: {}
      meta: {}
    type: object
  hexagonal-go-backend_internal_platform_http_response.ErrorBody:
    properties:
      code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  hexagonal-go-backend_internal_platform_http_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorBody'
      request_id:
        type: string
    type: object
  internal_modules_courses_delivery_http_v1.CreateActivityRequest:
    properties:
      is_required:
        type: boolean
      resources:
        items:
          $ref: '#/definitions/internal_modules_courses_delivery_http_v1.CreateResourceRequest'
        type: array
      sequence_order:
        type: integer
      title:
        type: string
      type:
        type: string
    required:
    - title
    - type
    type: object
  internal_modules_courses_delivery_http_v1.CreateCategoryRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      parent_id:
        type: string
      tenant_id:
        type: string
    required:
    - code
    - name
    - tenant_id
    type: object
  internal_modules_courses_delivery_http_v1.CreateCourseRequest:
    properties:
      modules:
        items:
          $ref: '#/definitions/internal_modules_courses_delivery_http_v1.CreateModuleRequest'
        type: array
      syllabus:
        $ref: '#/definitions/internal_modules_courses_delivery_http_v1.SetSyllabusRequest'
      template:
        $ref: '#/definitions/internal_modules_courses_delivery_http_v1.CreateTemplateRequest'
      version:
        $ref: '#/definitions/internal_modules_courses_delivery_http_v1.CreateVersionRequest'
    required:
    - template
    - version
    type: object
  internal_modules_courses_delivery_http_v1.CreateLessonRequest:
    properties:
      activities:
        items:
          $ref: '#/definitions/internal_modules_courses_delivery_http_v1.CreateActivityRequest'
        type: array
      sequence_order:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  internal_modules_courses_delivery_http_v1.CreateModuleRequest:
    properties:
      lessons:
        items:
          $ref: '#/definitions/internal_modules_courses_delivery_http_v1.CreateLessonRequest'
        type: array
      sequence_order:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  internal_modules_courses_delivery_http_v1.CreateResourceRequest:
    properties:
      metadata:
        additionalProperties: {}
        type: object
      mime_type:
        type: string
      name:
        type: string
      position:
        type: integer
      size_bytes:
        minimum: 0
        type: integer
      type:
        type: string
      url_storage_key:
        type: string
    required:
    - name
    - type
    - url_storage_key
    type: object
  internal_modules_courses_delivery_http_v1.CreateTemplateRequest:
    properties:
      code:
        type: string
      description:
        type: string
      title:
        type: string
    required:
    - code
    - title
    type: object
  internal_modules_courses_delivery_http_v1.CreateVersionRequest:
    properties:
      estimated_hours:
        minimum: 0
        type: integer
      status:
        enum:
        - DRAFT
        - PUBLISHED
        - ARCHIVED
        type: string
      tag:
        type: string
    required:
    - status
    - tag
    type: object
  internal_modules_courses_delivery_http_v1.SetSyllabusRequest:
    properties:
      duration_hours:
        minimum: 0
        type: integer
      entry_profile:
        type: string
      exit_profile:
        type: string
      methodology:
        type: string
      objectives:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: soporte@empresa.com
    name: API Support
  description: API REST construida con Go, Gin y Arquitectura Hexagonal.
  title: Go Hexagonal API
  version: "1.0"
paths:
  /api/v1/courses:
    post:
      consumes:
      - application/json
      parameters:
      - description: Course
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_courses_delivery_http_v1.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: Create course with its full content tree
      tags:
      - courses
  /api/v1/courses/{id}:
    get:
      parameters:
      - description: Version ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: Get course by version ID with its full content tree
      tags:
      - courses
  /api/v1/courses/categories:
    get:
      parameters:
      - description: Tenant ID
        in: query
        name: tenant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: List course categories
      tags:
      - courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Category
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_courses_delivery_http_v1.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: Create course category
      tags:
      - courses
  /api/v1/courses/templates:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: List course templates
      tags:
      - courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Template
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_courses_delivery_http_v1.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: Create course template
      tags:
      - courses
  /api/v1/courses/templates/{id}:
    get:
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: Get course template
      tags:
      - courses
  /api/v1/courses/templates/{id}/versions:
    get:
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: List course versions
      tags:
      - courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Version
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_courses_delivery_http_v1.CreateVersionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: Create course version
      tags:
      - courses
  /api/v1/courses/versions/{id}/syllabus:
    get:
      parameters:
      - description: Version ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: Get course syllabus
      tags:
      - courses
    put:
      consumes:
      - application/json
      parameters:
      - description: Version ID
        in: path
        name: id
        required: true
        type: string
      - description: Syllabus
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_courses_delivery_http_v1.SetSyllabusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/hexagonal-go-backend_internal_platform_http_response.ErrorResponse'
      summary: Set course syllabus
      tags:
      - courses
securityDefinitions:
  BearerAuth:
    description: Escribe "Bearer " seguido de tu token JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"